+++ CHANGELOG.md
@@ -10,10 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Send last filters update time during issue reporting [#3055].
+- Blocked by `$document` rules blocking page for MV3.
 
 ### Changed
 
-- Updated Safebrowsing and blocked by rules blocking pages.
+- Updated blocking pages for MV2 — Safebrowsing and blocked by `$document` rules.
 
 ### Removed